Find new command name for lowered or raised sections.
Don't split at node when splitting at chapter or section...

Handle sectioning commands without argument.  Make sure that no <a >
element is output if the text or the name and href are empty.

Add a default filename for sections too, using the normalized
transliterated section name.  Use it if NODE_FILENAMES is set.


git-svn-id: svn://svn.sv.gnu.org/texinfo/trunk@3904 39fee189-59d7-47db-b5d4-205258b72aed